Clip moved to official location
